# swagger/Datasheet_svc.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Datasheet Service
  description: This API  allows the consumer to perform several operations in a table matrix.
schemes:
  - http
host: localhost:3001
basePath: /
paths:
  /Users:
    post:
      tags:
        - Users
      description: Create a new item in the resource User.
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - name: user
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        201:
          description: User successfully created.
          schema:
            $ref: '#/definitions/UserResponse'
        400:
          description: Bad request. Check the definition documentation.
    get:
      tags:
        - Users
      description: Get all available registered users.
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      responses:
        200:
          description: User(s) was/were found.
          schema:
            type: array
            items: {
              "$ref": "#/definitions/SimpleUser"
            }
        400:
          description: Bad request. Check the definition documentation.
...
  /Users/{UserID}/Macros:
    post:
      tags:
        - Macros
      description: Creates a macro
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - name: UserID
          in: path
          required: true
          type: string
        - name: macro
          in: body
          required: true
          schema:
            $ref: '#/definitions/MacroRequest'
      responses:
        201:
          description: Macro was successfully created.
          schema:
              $ref: '#/definitions/Macro'
        400:
            description: Bad request. Check the definition documentation.
    get:
      tags:
        - Macros
      description: Gets all existent macros of current user.
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - name: UserID
          in: path
          required: true
          type: string
      responses:
        200:
          description: Macro(s) was/were found.
          schema:
            type: array
            items: {
              "$ref": "#/definitions/Macro"
            }
        400:
          description: Bad request. Check the definition documentation.
        404:
          description: Macro(s) was/were not found.
  /Users/{UserID}/Macros/{MacroID}:
    put:
      tags:
        - Macros
      description: Update a macro.
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - name: UserID
          in: path
          required: true
          type: string
        - name: MacroID
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Macro'
      responses:
        200:
          description: Macro successfully updated
        400:
          description: Bad request. Check the definition documentation.
        404:
          description: Macro not found.
...
  /Users/{UserID}/Datasets/{DatasetID}/Macros?MacroID=MacroID:
    post:
      tags:
        - Stats
      description: Calculate macro of entire dataset
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
       - name: UserID
         in: path
         required: true
         type: string
       - name: DatasetID
         in: path
         required: true
         type: string
       - name: MacroID
         in: query
         required: true
         type: string
      responses:
        202:
          description: Accepted.
          schema:
            $ref: "#/definitions/MacroResult"
        400:
          description: Bad request. Check the definition documentation.
...
definitions:
  Macro:
    type : object
    properties:
      idMacro:
        type: integer
      username:
        type: string
      items:
        type: array
        items: {
            "$ref": "#/definitions/Item"
          }
      description:
        type: string
  MacroRequest:
    type : object
    properties:
      items:
        type: array
        items: {
            "$ref": "#/definitions/Item"
          }
      description:
        type: string
  MacroResult:
    type : object
    properties:
      result_url:
        type: integer
  Item:
    type : object
    properties:
      stat_id:
        type: string
      transf_id:
        type: string
      value:
        type: integer
...
  SimpleUser:
    type: object
    properties:
      username:
        type: string
        description: Unique identifier as nickname. e.g. user123
      full name:
        type: string
        description: Your full name.
  User:
    type: object
    properties:
      username:
        type: string
        description: Unique identifier as nickname. e.g. user123
      full name:
        type: string
        description: Your full name.
      password:
        type: string
        description: Your password code.
...
  UserResponse:
    type: object
    properties:
      username:
        type: string
        description: Your full name confirming that you logged with sucess.
...